Crate tiny_dynamo

📦 🤏

tiny dynamo

A tinier, simpler, key-value focused interface for AWS DynamoDB

Tiny what now?

Amazon DynamoDB is a key-value and document database that delivers single-digit millisecond performance at any scale.

This quote comes directly from the Amazon DynamoDB docs. This has some implications that are less than ideal for very simple key-value applications. It can be overly complicated and sometimes daunting for the uninitiated to say the least.

Tiny Dynamo aims to leverge the useful parts by exposing a much simpler get/set API you might expect from a key-value interface.

Usage

use std::{env, error::Error};
use tiny_dynamo::{reqwest_requests::Reqwest, Credentials, TableInfo, DB};

fn main() -> Result<(), Box<dyn Error>> {
    let db = DB::new(
        Credentials::new(
            env::var("AWS_ACCESS_KEY_ID")?,
            env::var("AWS_SECRET_ACCESS_KEY")?,
        ),
        TableInfo::new(
            "key-attr-name",
            "value-attr-name",
            "table-name",
            "us-east-1".parse()?,
            None
        ),
        Reqwest::new(),
    );

    println!("{:#?}", db.set("foo", "bar")?);
    println!("{:#?}", db.get("foo")?);

    Ok(())
}

A few notable differences when comparing to traditional DynamoDB clients is that this client assumes a single table, a very common case, so you configure your client with that table name so you don’t need to redundantly provided it with each request.

You will also find the interface is reduced to get(key) set(key,value). This is intentional as this client is primarily focused on being a better fit for simple key-value applications.

Features

Tiny

Tiny Dynamo avoids packing carry-on luggage for anything you don’t explicitly need for a simple key-value application. This includes an entire sdk and transient line of dependencies. This allows it to fit more easily into smaller spaces and to deliver on Rust’s zero cost promise of not paying for what you don’t use.

Simpler Data Modeling

A common laborious activity with DynamoDB applications figuring our your application’s data model first and then translating that to a DynamoDB’s key space design and catalog of item attribute types. This is fine and expected for applications that require more advanced access patterns. For simple key-value applications, this is just tax. Tiny DynamoDB assumes a key-value data model. How you serilize your value is up to you.

Just the data plane

You can think of the DynamoDB API in terms of two planes: The data plane, where you typically spend all your time in 99% of application cases, and the control plane, an API for provisioning the resources that will store your data. Combining these makes its surface area arbitrarily larger that it needs to be. Tiny Dynamo focuses on exposing just the data plane to retain a smaller surface area to learn.

Sans I/O

Tiny Dynamo takes a sans I/O library approach. By default it defines a Requests trait which allows for any I/O library to implement how requests are transfered over the wire by provides none without an implicit feature toggled on

Reqwest

the reqwest feature provides a reqwest_requests::Reqwest backend for sending requests, currently using a blocking client. An async feature is planned for the future

[dependencies]
tiny_dynamo = { version = "0.1", features = ["reqwest"]}

Fastly

The fastly feature provides a fastly_requests::Fastly backend for sending requests suitable for Fastlys Compute@Edge platform

[dependencies]
tiny_dynamo = { version = "0.1", features = ["fastly"]}

BYOIO

If you would like to bring your own IO implementation you can define an implementation for a custom type

use tiny_dynamo::{Request, Requests};
use std::error::Error;

struct CustomIO;

impl Requests for CustomIO {
  fn send(&self, signed: Request) -> Result<(u16, String), Box<dyn Error>> {
    Ok(
      (200,"...".into())
    )
  }
}

Doug Tangren (softprops) 2021

Structs

Const
Provides a Requests implementation for a constantized response.

Credentials
A set of AWS credentials to authenticate requests with

DB
The central client interface applications will work with

TableInfo
Information about your target AWS DynamoDB table

Enums

Region
A list of AWS Regions supported by DynamoDB

Traits

Requests
A trait to implement the behavior for sending requests, often your “IO” layer

Type Definitions

Request
A type alias for http::RequestVec<u8>